--- src/components/Blog/BlogArticlesCollections/PaginationSkeleton.tsx
@@ -1,20 +1,0 @@
-import { Skeleton } from '@mui/material';
-
-import { PaginationContainer } from './Pagination.style';
-
-export const BlogArticlesBoardPaginationSkeleton = () => {
-  return (
-    <PaginationContainer>
-      {Array.from({ length: 4 }).map((_, index) => {
-        return (
-          <Skeleton
-            variant="circular"
-            width="40"
-            height="40"
-            key={`pagination-skeleton-${index}`}
-          />
-        );
-      })}
-    </PaginationContainer>
-  );
-};
